+++ docs/Todo.md
@@ -1,14 +1,5 @@
 # Todo
 
-* remove jquery (https://insidegovuk.blog.gov.uk/2022/08/11/how-and-why-we-removed-jquery-from-gov-uk/)
-  ^^ moj uses it
-
-* apply tweaks to setCookie from cookie_consent.ts
-
-* tidy up js files everywhere
-
-* take out typescript? more hassle than it's worth (especially with all the plain js we've borrowed)?
-
 * add js tests for govuk-design-system sourced js (https://github.com/alphagov/govuk-design-system)
 
 * cost : api needs to add new field and use that for the every {new_field} bit
