+++ src/FamilyHubs.ServiceDirectory.Web/Pages/Shared/_Layout.cshtml
@@ -72,7 +72,7 @@
     </script>
     
 @*todo: get optimization into this pr*@
-    <script src="~/js/scripts/app.js" type="module" asp-append-version="true"></script>
+    <script src="~/js/app.js" type="module" asp-append-version="true"></script>
 @*    <script src="~/js/components/cookie-banner.js" type="module" asp-append-version="true"></script>*@
     @await RenderSectionAsync("Scripts", required: false)
