+++ src/FamilyHubs.ServiceDirectory.Web/scripts/app.ts
@@ -6,10 +6,10 @@ declare global {
     }
 }
 
-import CookieBanner from '../wwwroot/js/components/cookie-banner.js'
-import { getConsentCookie, isValidConsentCookie } from '../wwwroot/js/components/cookie-functions.js'
-import Analytics from '../wwwroot/js/components/analytics.js'
-import CookiesPage from '../wwwroot/js/components/cookies-page.js'
+import CookieBanner from './components/cookie-banner.js'
+import { getConsentCookie, isValidConsentCookie } from './components/cookie-functions.js'
+import Analytics from './components/analytics.js'
+import CookiesPage from './components/cookies-page.js'
 
 // Initialise cookie banner
 var $cookieBanner = document.querySelector('[data-module="govuk-cookie-banner"]')
