+++ src/FamilyHubs.ServiceDirectory.Web/scripts/components/analytics.js
@@ -1,6 +1,8 @@
-//todo: incorporate nonce support to this from fts
+﻿//todo: incorporate nonce support to this from fts
+
 export default function loadAnalytics() {
     // no-op until we add ga later
+
     //if (!window.ga || !window.ga.loaded) {
     //    // Load gtm script
     //    // Script based on snippet at https://developers.google.com/tag-manager/quickstart
@@ -10,12 +12,13 @@ export default function loadAnalytics() {
     //            'gtm.start': new Date().getTime(),
     //            'event': 'gtm.js'
     //        })
+
     //        var j = d.createElement(s)
     //        var dl = l !== 'dataLayer' ? '&l=' + l : ''
+
     //        j.async = true
     //        j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl
     //        document.head.appendChild(j)
     //    })(window, document, 'script', 'dataLayer', 'GTM-53XG2JT')
     //}
 }
-//# sourceMappingURL=analytics.js.map
