+++ src/FamilyHubs.ServiceDirectory.Web/wwwroot/js/app.js
@@ -1,8 +1,8 @@
 // js components have been snaffled from https://github.com/alphagov/govuk-design-system/blob/main/src/javascripts/components/
-import CookieBanner from '../wwwroot/js/components/cookie-banner.js';
-import { getConsentCookie, isValidConsentCookie } from '../wwwroot/js/components/cookie-functions.js';
-import Analytics from '../wwwroot/js/components/analytics.js';
-import CookiesPage from '../wwwroot/js/components/cookies-page.js';
+import CookieBanner from './components/cookie-banner.js';
+import { getConsentCookie, isValidConsentCookie } from './components/cookie-functions.js';
+import Analytics from './components/analytics.js';
+import CookiesPage from './components/cookies-page.js';
 // Initialise cookie banner
 var $cookieBanner = document.querySelector('[data-module="govuk-cookie-banner"]');
 new CookieBanner($cookieBanner).init();
